projekt/backend/requirements.txt
Flask
Flask-CORS
python-keycloak
psycopg2
python-dotenv
Authlib
requests